pyproject.toml
[tool.poetry]
name = "repl_python3_beta_StudyNow-Website"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.0.1"
pymongo = "^3.11.4"
dnspython = "^2.1.0"
pycryptodomex = "^3.10.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
